+++ source/headers.h
@@ -1,8 +1,8 @@
 #include "raylib.h"
 
 //interface
-void HUD(short isNights, short pointCounter, short ideyaCounter, short currentTime, short overallTime, int screenWidth, int screenHeight);
-void GraphicalMenu(int screenWidth, int screenHeight, short *gameState);
+void HUD(int isNights, int pointCounter, int ideyaCounter, int currentTime, int overallTime, int screenWidth, int screenHeight);
+void GraphicalMenu(int screenWidth, int screenHeight, int *gameState);
 
 // world
 Model LoadWorldModel(char* modelPath);
@@ -11,9 +11,9 @@ void WorldMain(char* modelPath);
 
 // character
 void DrawCharacter(Vector3 initialPosition, char* modelPath);
-void MoveCharacter3D(Vector3 *position, Camera3D *camera, short isNights);
+void MoveCharacter3D(Vector3 *position, Camera3D *camera, int isNights);
 void JumpCharacter3D(Vector3 *position, Camera3D *camera);
 
 // camera
-void RotateCamera(Camera3D *camera, Vector3 *characterPosition, short *cameraAngle, short isNights);
-void SwitchCameraTo2dot5D(Camera3D *camera, short isNights, short *cameraAngle);
+void RotateCamera(Camera3D *camera, Vector3 *characterPosition, int *cameraAngle, int isNights);
+void SwitchCameraTo2dot5D(Camera3D *camera, int isNights, int *cameraAngle);
